% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMissingColsAndFilIWith0.R
\name{AddMissingColsAndFillWith0}
\alias{AddMissingColsAndFillWith0}
\title{AddMissingColsAndFillWith0}
\usage{
AddMissingColsAndFillWith0(df, ALL_COLS)
}
\arguments{
\item{df}{dataframe from JPK file, for now is: SprzedazWiersz, ZakupyWiersz, FakturaWiersz, Faktura}

\item{ALL_COLS}{character vector with list of all cols of full JPK inside table, for now is: SprzedazWiersz, ZakupyWiersz, FakturaWiersz, Faktura}
}
\value{
df filled with all missing columns.
}
\description{
AddMissingColsAndFillWith0
}
