% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIG_XML_TO_SQLITE.R
\name{BIG_XML_TO_SQLITE}
\alias{BIG_XML_TO_SQLITE}
\title{Title}
\usage{
BIG_XML_TO_SQLITE(file_xml = "", file_SQL = "Plik_JPK.sqlite",
  bufor_size = 1000)
}
\arguments{
\item{bufor_size}{}
}
\description{
Title
}
