% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCharColsToNum.R
\name{convertCharColsToNum}
\alias{convertCharColsToNum}
\title{convertCharColsToNum}
\usage{
convertCharColsToNum(df, colsToConv, guess = F)
}
\arguments{
\item{df}{df of wchich cols you want to convert}

\item{colsToConv}{if guess is False, you have to pass list of cols by names like c("K_10", "K_15")}

\item{guess}{logical if guess is True, then function will guess automatically which columns have only numbers and have to be converted to numeric}
}
\value{
returns tibble (kind of data.frame) where cols which are only numeric (but class character), are converted to numeric
}
\description{
convertCharColsToNum
}
\seealso{
\code{\link{as.tibble}}
}
