% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToSQLite.R
\name{csvToSQLite}
\alias{csvToSQLite}
\title{CSV to sqlite";"}
\usage{
csvToSQLite(pathOfCSVfiles = "", SQLiteConnection = myDB,
  jpk_type = NA)
}
\arguments{
\item{SQLiteConnection}{conn to SQLite file}

\item{jpk_type}{type of jpk for example "JPK_VAT", "JPK_FA"}

\item{file_xml}{location of xml file}

\item{export_dir}{directory where csv files will be written}

\item{bufor_size}{size of one time write to connection}
}
\value{
csv files of with lines of jpk tables and sqlconnection filled with this files
}
\description{
CSV to sqlite";"
}
