% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doALLtests.R
\name{doALLtests}
\alias{doALLtests}
\title{doALLtests}
\usage{
doALLtests(SQLite_file = NA, export_dir, customFileName,
  WybraneTesty = 1:100)
}
\arguments{
\item{testReturn}{must be list which is returned after one of test fun test001, 002 ...}
}
\value{
exported file
}
\description{
bla bla bla
}
