% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIG_XML_TO_SQLITE.R
\name{exportJPKtoSQLite}
\alias{exportJPKtoSQLite}
\title{exportJPKtoSQLite is doing parsing JPK elemens for example from "Faktura" table of JPK_FA file single record is between this characters "<Faktura typ=\"G\">","</Faktura>",
function generate SQLite file with tables with records of JPK file.}
\usage{
exportJPKtoSQLite(file_xml = "", file_SQL = "Plik_JPK.sqlite",
  bufor_size = 1000)
}
\arguments{
\item{bufor_size}{}
}
\description{
exportJPKtoSQLite is doing parsing JPK elemens for example from "Faktura" table of JPK_FA file single record is between this characters "<Faktura typ=\"G\">","</Faktura>",
function generate SQLite file with tables with records of JPK file.
}
