% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getControlSum.R
\name{getControlSums}
\alias{getControlSums}
\title{getControlSums}
\usage{
getControlSums(pointer_of_file_xml, xPath)
}
\arguments{
\item{pointer_of_file_xml}{Pointer to the xml file which was read with xml2::read_xml function}

\item{xPath}{xpath for node for example "//d1:FakturaCtrl"}
}
\value{
numeric tibble with control sums
}
\description{
getControlSums
}
