% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getControlSums_old.R
\name{getControlSums_old}
\alias{getControlSums_old}
\title{getControlSums_old}
\usage{
getControlSums_old(pointer_of_file_xml, xPath)
}
\arguments{
\item{pointer_of_file_xml}{Pointer to the xml file which was read with xml2::read_xml function}

\item{xPath}{xpath for node for example "//d1:FakturaCtrl"}
}
\value{
numeric tibble with control sums
}
\description{
getControlSums_old
}
