% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGapsReport.R
\name{getGapsReport}
\alias{getGapsReport}
\title{getGapsReport
Sluzy do zweryfikowania czy sa luki w fakturach
#'}
\usage{
getGapsReport(patternToMatch, data)
}
\arguments{
\item{data}{must be vector with invoice numbers}
}
\value{
data frames with gaps
}
\description{
getGapsReport
Sluzy do zweryfikowania czy sa luki w fakturach
#'
}
