% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPointerOfXML.r
\name{getPointerOfXML}
\alias{getPointerOfXML}
\title{getPointerOfXML xml2::read_xml(x = file_xml)}
\usage{
getPointerOfXML(file_xml = "")
}
\arguments{
\item{file_xml}{}
}
\value{
pointer to xml file
}
\description{
getPointerOfXML xml2::read_xml(x = file_xml)
}
