% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_xmlToCSV.R
\name{GO_xmlToCSV}
\alias{GO_xmlToCSV}
\title{xml to CSV"}
\usage{
GO_xmlToCSV(file_xml = "", export_dir = "", jpk_type = JPK_TYPE,
  uniqueNamePart = stringr::str_remove_all(string = Sys.time(), pattern =
  "[^0-9]"), pathToBinary = Sys.getenv(x = "JPKr_GO_BINARY"))
}
\arguments{
\item{file_xml}{location of xml file}

\item{export_dir}{directory where csv files will be written}

\item{jpk_type}{type of jpk for example "JPK_VAT", "JPK_FA"}

\item{SQLiteConnection}{conn to SQLite file}

\item{bufor_size}{size of one time write to connection}
}
\value{
csv files of with lines of jpk tables and sqlconnection filled with this files
}
\description{
xml to CSV"
}
