% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_xmlToSQLite.R
\name{GO_xmlToSQLite}
\alias{GO_xmlToSQLite}
\title{Title}
\usage{
GO_xmlToSQLite(file_xml = "", SQLiteConnection = myDB, export_dir = "",
  bufor_size = 10000, jpk_type = JPK_TYPE)
}
\arguments{
\item{jpk_type}{}
}
\description{
Title
}
