% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linesToSQLite.R
\name{linesToSQLite}
\alias{linesToSQLite}
\title{linesToSQLite}
\usage{
linesToSQLite(linesOfJPK = JPK_UNKNOWN_text, bufor_size = 1000,
  SQLiteConnection, table_name, record_start, record_end,
  removeStringfromColnames = "^(Faktura\\\\.)", colnamesList)
}
\arguments{
\item{bufor_size}{size of bufor}

\item{SQLiteConnection}{sqlite connection}

\item{table_name}{name of table to export data}

\item{record_start}{record start}

\item{record_end}{record end}

\item{removeStringfromColnames}{record end}

\item{colnamesList}{record end}

\item{file_xml}{path to JPK xml file}
}
\description{
linesToSQLite
}
