% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInvoiceNumbers.R
\name{parseInvoiceNumbers}
\alias{parseInvoiceNumbers}
\title{parseInvoiceNumbers
Sluzy do zweryfikowania czy sa luki w fakturach
#'}
\usage{
parseInvoiceNumbers(data)
}
\arguments{
\item{data}{must be vector with invoice numbers}
}
\value{
data frames with gaps
}
\description{
parseInvoiceNumbers
Sluzy do zweryfikowania czy sa luki w fakturach
#'
}
