% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJPKtoMemory.R
\name{readJPKtoMemory}
\alias{readJPKtoMemory}
\title{readJPKtoMemory reads file as list no parsing is done.}
\usage{
readJPKtoMemory(file_xml = "")
}
\arguments{
\item{file_xml}{}
}
\value{
only non-parsed list
}
\description{
readJPKtoMemory reads file as list no parsing is done.
}
