% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmCharNULLfromDF.R
\name{rmCharNULLfromDF}
\alias{rmCharNULLfromDF}
\title{rmCharNULLfromDF In JPK files NA values are like character 'NULL'. This funcion replaces 'NULL's' with '0'}
\usage{
rmCharNULLfromDF(df)
}
\arguments{
\item{df}{data frame JPK record}
}
\value{
df with replaced all character NULL's to 0
}
\description{
rmCharNULLfromDF In JPK files NA values are like character 'NULL'. This funcion replaces 'NULL's' with '0'
}
