% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteToExcel.R
\name{SQLiteToExcel}
\alias{SQLiteToExcel}
\title{Export tabel pliku SQLite do Excela (xlsx)}
\usage{
SQLiteToExcel(file_SQL = NA, file_excel = NA)
}
\arguments{
\item{file_SQL}{path to sqlitefile}

\item{file_excel}{where you want o save SQLite file}
}
\value{
xlsx file with JPK tables
}
\description{
Dodać
}
