% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test001.r
\name{test001}
\alias{test001}
\title{test001 Służy do obliczenia sum kontrolnych.}
\usage{
test001(data)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
a tibble with some metric: sum, median, mean and etc..
}
\description{
test001 Służy do obliczenia sum kontrolnych.
}
