% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test002.r
\name{test002}
\alias{test002}
\title{test002 Służy do obliczenia sum kontrolnych.}
\usage{
test002(data)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
a tibble with some metric: sum, median, mean and etc..
}
\description{
test002 Służy do obliczenia sum kontrolnych.
}
