% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test003.R
\name{test003}
\alias{test003}
\title{test003 Sluzy do zweryfikowania koncentracji dla danego NIP z pliku JPK}
\usage{
test003(data = NA, tableName = NA)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
return tibble with summarise of values, counts, max, min, of documents for one NIP
}
\description{
test003 Sluzy do zweryfikowania koncentracji dla danego NIP z pliku JPK
}
