% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test004.r
\name{test004}
\alias{test004}
\title{test004
Sluzy do podsumowania kontrahentow w podziale na polski i zagraniczny NIP
T/F czy polski nip, NIP kontrahenta, nazwa kontrahenta, wartosc dokumentow, liczba dokumentow}
\usage{
test004(data = NA, tableName = NA)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
return tibble with summarise of values, counts, max, min, of documents for one NIP
}
\description{
test004
Sluzy do podsumowania kontrahentow w podziale na polski i zagraniczny NIP
T/F czy polski nip, NIP kontrahenta, nazwa kontrahenta, wartosc dokumentow, liczba dokumentow
}
