% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test005.R
\name{test005}
\alias{test005}
\title{parseInvoiceNumbers
Sluzy do zweryfikowania czy sa luki w fakturach
#'}
\usage{
test005(data, tabName = NA)
}
\arguments{
\item{data}{must be vector with invoice numbers}
}
\value{
list with reports
}
\description{
parseInvoiceNumbers
Sluzy do zweryfikowania czy sa luki w fakturach
#'
}
