% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test006.R
\name{test006}
\alias{test006}
\title{test005
Służy do zweryfikowania czy podatki zostały właściwie obliczone}
\usage{
test006(data = NA, tableName = NA)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
return tibble with summarise of values, counts, max, min, of documents for one NIP
}
\description{
test005
Służy do zweryfikowania czy podatki zostały właściwie obliczone
}
