% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_ValidateControlSums.R
\name{Test_ValidateControlSums}
\alias{Test_ValidateControlSums}
\title{Test_ValidateControlSums test if two values are equal with '==' and returns error, or warning. This depends form onlyWarning param}
\usage{
Test_ValidateControlSums(V1, V2, msg, onlyWarning = F)
}
\arguments{
\item{V1}{First value - scalar}

\item{V2}{Second value - scalar}

\item{msg}{Message formated to sprintf function}

\item{onlyWarning}{param to choose if user want to only warn or error. Warning allows funcion to finish, error stops function}
}
\value{
error or warning with msg
}
\description{
Test_ValidateControlSums test if two values are equal with '==' and returns error, or warning. This depends form onlyWarning param
}
\seealso{
\code{\link{sprintf}}
}
