% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testExport.R
\name{testExport}
\alias{testExport}
\title{test001 Służy do obliczenia sum kontrolnych.}
\usage{
testExport(testReturn, export_dir)
}
\arguments{
\item{testReturn}{must be list which is returned after one of test fun test001, 002 ...}
}
\value{
exported file
}
\description{
test001 Służy do obliczenia sum kontrolnych.
}
