% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationOfPolishNIP.R
\name{validationOfPolishNIP}
\alias{validationOfPolishNIP}
\title{Title function checks if NIP number 'Polish company indetyfication number' is from Poland 'TRUE' or not 'FALSE'}
\usage{
validationOfPolishNIP(NIP)
}
\arguments{
\item{NIP}{}
}
\value{
logical
}
\description{
Title function checks if NIP number 'Polish company indetyfication number' is from Poland 'TRUE' or not 'FALSE'
}
