% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endTimePrediction.R
\name{endTimePrediction}
\alias{endTimePrediction}
\title{endTimePrediction Służy do obliczenia ile pozostało czasu przy wczytaniu pliki.}
\usage{
endTimePrediction(start, index, lastIndex, TimeFromStart = F)
}
\arguments{
\item{start}{numericPositict}

\item{index}{numeric index from loop like 1,2,3, ... n}

\item{lastIndex}{last index. end of loop}

\item{TimeFromStart}{T/F For measuring not how much time left, but how much time passed}
}
\value{
character string with prediction of time left
}
\description{
endTimePrediction Służy do obliczenia ile pozostało czasu przy wczytaniu pliki.
}
