% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessJpkType.R
\name{getJpkType}
\alias{getJpkType}
\title{getJpkType}
\usage{
getJpkType(pointer_of_file_xml)
}
\arguments{
\item{pointer_of_file_xml}{}
}
\value{
type of JPK file like 'JPK_FA', 'JPK_VAT3'
}
\description{
getJpkType
}
