% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test002.r
\name{test002}
\alias{test002}
\title{test002 Sluzy do zweryfikowania sum kontrolnych z pliku JPK}
\usage{
test002(data = NA, controlSumsFromFile = NA, jpk_type = NA)
}
\arguments{
\item{data}{must be converted to numeric data.frame with JPK FA, or JPK_VAT}
}
\value{
return list with 3 elements.

 1. Logical TRUE, or FALSE depends on if control sums computed vs control sums from JPK file are equal,

 2. Control sums from JPK file,

 3. Control sums computed from 'data'
}
\description{
test002 Sluzy do zweryfikowania sum kontrolnych z pliku JPK
}
